+++ transactions/pkg/services/response.go
@@ -4,14 +4,14 @@ import "github.com/maslow123/transactions/pkg/pb"
 
 func genericCreateTransactionResponse(statusCode int, errorMessage string) (*pb.CreateTransactionResponse, error) {
 	return &pb.CreateTransactionResponse{
-		Status: int64(statusCode),
+		Status: int32(statusCode),
 		Error:  errorMessage,
 	}, nil
 }
 
 func genericGetTransactionListByUserResponse(statusCode int, errorMessage string) (*pb.GetTransactionListResponse, error) {
 	return &pb.GetTransactionListResponse{
-		Status: int64(statusCode),
+		Status: int32(statusCode),
 		Error:  errorMessage,
 	}, nil
 }
